requirements.txt
pandas
numpy
python-dotenv
redis